+++ fixtures/remix-1.18.0/express/metronome.config.js
@@ -5,4 +5,8 @@ export default {
   ignoredRoutes: [],
   ignoredPathnames: ["/healthcheck"],
   ignoreHeadMethod: false,
+  doNotTrack: (request) => {
+    // console.log("doNotTrack", { request });
+    return true;
+  },
 };