+++ packages/config/src/index.ts
@@ -1,2 +1,3 @@
 export { MetronomeConfigHandler } from "./MetronomeConfigHandler";
 export type { MetronomeConfig } from "./types";
+export { findConfigFile } from "./findConfigFile";