+++ packages/config/src/metronome.config.js
@@ -5,4 +5,8 @@ module.exports = {
   ignoredRoutes: [],
   ignoredPathnames: ["/healthcheck"],
   ignoreHeadMethod: true,
+  doNotTrack: (request) => {
+    // Track all events
+    return false;
+  },
 };