+++ packages/config/src/types.ts
@@ -2,4 +2,5 @@ export interface MetronomeConfig {
   ignoredRoutes?: (string | RegExp)[];
   ignoredPathnames?: (string | RegExp)[];
   ignoreHeadMethod?: boolean;
+  doNotTrack?: (request: Request) => Promise<boolean | undefined | null>;
 }