+++ packages/express/src/index.ts
@@ -1,3 +1,4 @@
 export { createRequestHandler } from "./createRequestHandler";
 export { wrapRequestHandlerWithMetronome } from "./wrapRequestHandlerWithMetronome";
 export type { MetronomeConfig } from "@metronome-sh/config";
+export { compose } from "@metronome-sh/runtime";