+++ packages/react/src/index.ts
@@ -1,1 +1,2 @@
 export { withMetronome } from "./withMetronome";
+export { compose } from "./utils/compose";