+++ packages/runtime/src/index.ts
@@ -1,9 +1,7 @@
-export { AbstractSpan, SpanName } from "./AbstractSpan";
 export { METRONOME_CONTEXT_KEY, METRONOME_VERSION } from "./constants";
 export { registerMetronome } from "./registerMetronome";
 export { reportRoute } from "./routes";
 
-export type { SpanEvent } from "./AbstractSpan";
 export type {
   ContextWithMetronome,
   GetLoadContextOptions,
@@ -25,3 +23,5 @@ export {
 } from "./events";
 
 export { EventExporter } from "./EventExporter";
+
+export { compose } from "./utils/compose";