+++ packages/node/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @metronome-sh/node
 
+## 7.1.0-next.4
+
+### Patch Changes
+
+- Fixed config file
+- Updated dependencies
+  - @metronome-sh/runtime@7.1.0-next.4
+
 ## 7.1.0-next.3
 
 ### Patch Changes