+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @metronome-sh/react
 
+## 7.1.0-next.4
+
+### Patch Changes
+
+- Fixed config file
+
 ## 7.1.0-next.3
 
 ### Patch Changes